Auto detect .mzXML files.
