--- chemclipse/plugins/org.eclipse.chemclipse.msd.converter.supplier.mzxml/src/org/eclipse/chemclipse/msd/converter/supplier/mzxml/converter/MagicNumberMatcher.java
@@ -1,36 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2016, 2018 Lablicate GmbH.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- * 
- * Contributors:
- * Dr. Philip Wenig - initial API and implementation
- *******************************************************************************/
-package org.eclipse.chemclipse.msd.converter.supplier.mzxml.converter;
-
-import java.io.File;
-
-import org.eclipse.chemclipse.converter.core.AbstractMagicNumberMatcher;
-import org.eclipse.chemclipse.converter.core.IMagicNumberMatcher;
-import org.eclipse.chemclipse.msd.converter.supplier.mzxml.internal.io.SpecificationValidator;
-
-public class MagicNumberMatcher extends AbstractMagicNumberMatcher implements IMagicNumberMatcher {
-
-	@Override
-	public boolean checkFileFormat(File file) {
-
-		boolean isValidFormat = false;
-		try {
-			file = SpecificationValidator.validateSpecification(file);
-			if(file.exists()) {
-				return true;
-			}
-		} catch(Exception e) {
-			// Print no exception.
-		}
-		return isValidFormat;
-	}
-}
